Measure also clone() to start a thread.
This was a bit tricky because uclibc doesn't allow to call
clone directly. So, I've copied their assembly code for clone
and adjusted it accordingly. This works but we can't call
arbitrary library functions with the thread because uclibc
assumes that pthreads are used. E.g. in printf it tries to
get the stacksize as some point via pthread-attributes.