Reworked disk usage on gem5.
We now put all data into rootfs/ to have all data in the root FS on hw.
At the same time, we build two disk images for gem5. One as the root
FS, containing everything, and one as the benchmark FS, only containing
rootfs/bench. At runtime, we overmount /bench to see the current
benchmark data, which we might have changed since the last warmup.

This way, we can do the warmup once with the root FS and afterwards
change the benchmark data (e.g., recompile an application) without
having to do the warmup again.