Simplified benchmark automation on gem5.
Previously, we controlled Linux by writing to the serial line and
reading the results from it. This was pretty complicated, error prone,
and non-deterministic.

For these reasons, I've now replaced it with a new approach that uses a
third file system with a generated file that specifies what to execute.
This FS is changed for every run and only mounted after the initial
checkpoint.